requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.40.0,<5
accelerate>=0.27.0
ultralytics>=8.0.0
opencv-python>=4.8.0
supervision>=0.20.0
Pillow>=10.0.0
numpy>=1.24.0
pyyaml>=6.0
python-dotenv>=1.0.0
tabulate>=0.9.0
tqdm>=4.65.0
einops>=0.7.0
tiktoken>=0.6.0
